refactor(memory): move recall out of the tool module
